engine: added time to ray & separate object of hitable

+++ src/config.h
@@ -10,4 +10,4 @@
 
 #define MAX_BOUNCE_DEPTH 64
 
-typedef double rt_float;
+typedef float rt_float;
